+++ LEGO1/lego/legoomni/include/helicopter.h
@@ -44,17 +44,16 @@ public:
 	MxU32 m_unk0x08; // 0x08
 };
 
+#include "mxgeometry/mxgeometry4d.h"
+
 // VTABLE: LEGO1 0x100d40f8
 // VTABLE: BETA10 0x101b9880
 // SIZE 0x230
 class Helicopter : public IslePathActor {
 public:
-	Helicopter();
-	~Helicopter() override; // vtable+0x00
-
 	// FUNCTION: LEGO1 0x10003070
 	// FUNCTION: BETA10 0x1002b300
-	const char* ClassName() const override // vtable+0x0c
+	virtual const char* ClassName() const override // vtable+0x0c
 	{
 		// STRING: LEGO1 0x100f0130
 		return "Helicopter";
@@ -62,10 +61,13 @@ public:
 
 	// FUNCTION: LEGO1 0x10003080
 	// FUNCTION: BETA10 0x1002b330
-	MxBool IsA(const char* p_name) const override // vtable+0x10
+	virtual MxBool IsA(const char* p_name) const override // vtable+0x10
 	{
 		return !strcmp(p_name, Helicopter::ClassName()) || IslePathActor::IsA(p_name);
 	}
+
+	Helicopter();
+	~Helicopter() override; // vtable+0x00
 
 	MxResult Create(MxDSAction& p_dsAction) override;                            // vtable+0x18
 	void Animate(float p_time) override;                                         // vtable+0x70
